First fibonacci done

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,7 +12,7 @@ int main(void)
 
 	int second = 1;
 
-	int sum, i;
+	long int sum, i;
 
 	for (i = 0; i < 50; i++)
 	{
